+++ src/routes/api/teams.routes.js
@@ -1,4 +1,4 @@
-const { getAll, getById, getByName, create, updateOne, deleteOne, getByCategory } = require('../../controllers/teams.controller');
+const { getAll, getById, getByName, create, updateOne, deleteOne, getByCategory, getOwnerById } = require('../../controllers/teams.controller');
 
 
 const router = require('express').Router();
@@ -7,10 +7,12 @@ router.get('/', getAll)
 router.get('/:teamId', getById)
 router.get('/name/:teamName', getByName)
 router.get('/category/:category', getByCategory)
+router.get('/owner/:teamId', getOwnerById)
 
 router.post('/create', create)
 router.put('/:teamId', updateOne)
 router.delete('/:teamId', deleteOne)
+
 
 
 module.exports = router
